+++ lib/models/reserva_model.dart
@@ -1,11 +1,11 @@
-class Reserva {
+class CitasReservadas {
   final String fechaHoraReserva;
 
-  Reserva({required this.fechaHoraReserva});
+  CitasReservadas({required this.fechaHoraReserva});
 
   // Método de fábrica para crear una instancia a partir de un Map
-  factory Reserva.fromMap(Map<String, dynamic> map) {
-    return Reserva(
+  factory CitasReservadas.fromMap(Map<String, dynamic> map) {
+    return CitasReservadas(
       fechaHoraReserva: map['FechaHora_Reserva'] ?? '', // Extrae el campo FechaHora_Reserva
     );
   }
