+++ pubspec.yaml
@@ -10,6 +10,7 @@ dependencies:
   flutter:
     sdk: flutter
   http: ^1.2.2  
+  provider: 6.1.2
 
 dev_dependencies:
   flutter_test:
